+++ manufacturing_reports/__manifest__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-#############################################################################
+###############################################################################
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>).
-#    Author: Javid A(<https://www.cybrosys.com>)
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Author: Javid A (odoo@cybrosys.com)
 #
 #    You can modify it under the terms of the GNU AFFERO
 #    GENERAL PUBLIC LICENSE (AGPL v3), Version 3.
@@ -18,13 +18,16 @@
 #    (AGPL v3) along with this program.
 #    If not, see <http://www.gnu.org/licenses/>.
 #
-#############################################################################
+###############################################################################
 {
     'name': 'Manufacturing Reports',
     'version': '15.0.1.0.0',
-    'summary': 'PDF & XLS Reports For Manufacturing Module',
-    'description': 'PDF & XLS reports for manufacturing module with advanced filters.',
     'category': 'Manufacturing',
+    'summary': 'PDF & XLS Reports For Manufacturing Module',
+    'description': """This module crafted by Cybrosys Technologies provides 
+                      an option to print PDF & XLS report for MRP systems, 
+                      also you can view the product image in the report. This 
+                      works well for large and small business MRP systems.""",
     'author': 'Cybrosys Techno Solutions',
     'website': "http://www.cybrosys.com",
     'company': 'Cybrosys Techno Solutions',
@@ -32,10 +35,10 @@
     'depends': ['base', 'mrp'],
     'data': [
         'security/ir.model.access.csv',
-        'wizards/mrp_wizard_view.xml',
-        'reports/mrp_report_template.xml',
-        'reports/mrp_report.xml',
-        'views/menu_items.xml',
+        'reports/mrp_report_templates.xml',
+        'reports/mrp_report_views.xml',
+        'wizard/mrp_report_views.xml',
+        'views/manufacturing_report_menus.xml',
     ],
     'assets': {
         'web.assets_backend': [