+++ manufacturing_reports/doc/RELEASE_NOTES.md
@@ -3,4 +3,9 @@
 #### 17.02.2023
 #### Version 15.0.1.0.0
 ##### ADD
-- Initial commit
+- Initial commit for Manufacturing Reports
+
+#### 10.07.2023
+#### Version 15.0.1.0.0
+#### UPDT
+- Bug Fix for the xlsx report not printing issue in Manufacturing Reports